Separate into "utils".
Addresses in issue #2.
